requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.2
django-cleanup==5.2.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
passlib==1.7.4
paypalrestsdk==1.13.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.1.0
python-decouple==3.5
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stripe==3.5.0
urllib3==1.26.12